Adding a doc playground
As the docs team is moving to RST, this commit adds some
basic RST structure that could be used for the eventual
tutorial.

Of note, it includes conditional output logic, so specifying

make libcloud

will invoke sphinx-build with -t libcloud, meaning sections
marked .. only:: libcloud in the RST will be built, while others
won't. This could be a pathway forward for multi-language
support in the tutorial.