Use Tox to build out RST files
This commit adds support for Tox, and creates a directory structure
similar to the docs of other projects, where the RST files are located
in doc/source and the artifacts are placed in doc/build.

Tox is useful because it will track the python & sphinx dependencies and
also is the toolchain of choice for other repositories that use RST
files for docs, such as the specs repositories.